+++ src/s2uc.py
@@ -19,19 +19,6 @@ def cli():
 linkprompt = "Please authenticate with Globus here"
 def get_client():
     return NativeAppAuthClient('4787c84e-9c55-4881-b941-cb6720cea11c')
-# TODO Create configuration subsystem instead of hardcoding CLIENT ID values
-def get_scope_id(s2cs):
-    scope_map={
-        "10.16.42.61": "c42c0dac-0a52-408e-a04f-5d31bfe0aef8",
-        "10.16.41.12": "26c25f3c-c4b7-4107-8a25-df96898a24fe",
-        "10.16.42.31": "c42c0dac-0a52-408e-a04f-5d31bfe0aef8",
-        "localhost": "c42c0dac-0a52-408e-a04f-5d31bfe0aef8",
-        "127.0.0.1": "c42c0dac-0a52-408e-a04f-5d31bfe0aef8"
-    }
-    ip = s2cs.split(":")[0]
-
-    return scope_map.get(ip, "")
-    ## When IP is not found error silently, maybe not desirable
 
 @cli.command()
 @click.option('--scope', default="c42c0dac-0a52-408e-a04f-5d31bfe0aef8")
@@ -48,8 +35,11 @@ def login(scope):
     """
     adapter = utils.storage_adapter()
     try:
-        tokens= utils.get_access_token(scope)
-        click.echo("You are already logged in, to try different credentials please log out!")
+        tokens = utils.get_access_token(scope)
+        if "INVALID_TOKEN" in tokens:
+            raise utils.UnauthorizedError
+        else:
+            click.echo("You are already logged in, to try different credentials please log out!")
         return
     except utils.UnauthorizedError:
         click.echo("To obtain token for the scope, please open the URL in your browser and follow the instructions")
@@ -75,7 +65,7 @@ def check_auth(ip, scope):
     Displays globus credentials for a given ip or scope.
     """
     if ip:
-        scope = get_scope_id(ip)
+        scope = utils.get_scope_id(ip)
     if scope:
         token = utils.get_access_token(scope)
         click.echo(f"Access Token for scope '{scope}': {token}")
@@ -100,162 +90,59 @@ def logout():
 #TODO update
 @cli.command()
 @click.argument('uid', type=str, required=True)
-@click.option('--producer-s2cs', default="localhost:5000")
-@click.option('--consumer-s2cs', default="localhost:6000")
+@click.option('--s2cs', default="localhost:5000")
 @utils.authorize
-def release(uid, producer_s2cs, consumer_s2cs, metadata=None):
-    for s2cs in [producer_s2cs, consumer_s2cs]:
-        try:
-            with grpc.insecure_channel(s2cs) as channel:
-                stub = scistream_pb2_grpc.ControlStub(channel)
-                msg = scistream_pb2.Release(uid=uid)
-                resp = stub.release(msg, metadata=metadata)
-                print("Release completed")
-        except Exception as e:
-            print(f"Error during release: {e}")
-
-@cli.command()
-@click.option('--num_conn', type=int, default=5)
-@click.option('--rate', type=int, default=10000)
-@click.option('--producer-s2cs', default="localhost:5000")
-@click.option('--producer-s2cs', default="localhost:6000")
-def request(num_conn, rate, producer_s2cs, consumer_s2cs):
-    with grpc.insecure_channel(producer_s2cs) as channel1, \
-      grpc.insecure_channel(consumer_s2cs) as channel2:
-        prod_stub = scistream_pb2_grpc.ControlStub(channel1)
-        cons_stub = scistream_pb2_grpc.ControlStub(channel2)
-        uid=str(uuid.uuid1())
-        with futures.ThreadPoolExecutor(max_workers=4) as executor:
-            prod_resp_future = executor.submit(client_request, prod_stub, uid, "PROD", num_conn, rate)
-            cons_resp_future = executor.submit(client_request, cons_stub, uid, "CONS", num_conn, rate)
-            time.sleep(0.5)  # Possible race condition between REQ and HELLO
-            producer_future = executor.submit(AppCtrl, uid, "PROD", producer_s2cs, utils.get_access_token())
-            consumer_future = executor.submit(AppCtrl, uid, "CONS", consumer_s2cs, utils.get_access_token())
-
-            prod_resp = prod_resp_future.result()
-            cons_resp = cons_resp_future.result()
-            producer = producer_future.result()
-            consumer = consumer_future.result()
-
-        print(prod_resp) ## Should this be printed?
-        prod_lstn = prod_resp.listeners
-        prod_app_lstn = prod_resp.prod_listeners
-        cons_lstn = cons_resp.listeners
-
-        print("Sending updated connection map information...")
-        print(uid)
-        update(prod_stub, uid, prod_resp.prod_listeners)
-        update(cons_stub, uid, prod_resp.listeners)
-
-@cli.command()
-@click.option('--num_conn', type=int, default=5)
-@click.option('--rate', type=int, default=10000)
-@click.option('--producer-s2cs', default="localhost:5000")
-@click.option('--consumer-s2cs', default="localhost:6000")
-def request3(num_conn, rate, producer_s2cs, consumer_s2cs):
-    with grpc.insecure_channel(producer_s2cs) as channel1, \
-      grpc.insecure_channel(consumer_s2cs) as channel2:
-        prod_stub = scistream_pb2_grpc.ControlStub(channel1)
-        cons_stub = scistream_pb2_grpc.ControlStub(channel2)
-        uid=str(uuid.uuid1())
-        with futures.ThreadPoolExecutor(max_workers=4) as executor:
-            click.echo("uid; role; s2cs; access_token")
-            click.echo(f"{uid} {producer_s2cs} {utils.get_access_token()} PROD")
-            click.echo(f"{uid} {consumer_s2cs} {utils.get_access_token()} CONS")
-            click.echo("waiting for hello message")
-            prod_resp_future = executor.submit(client_request, prod_stub, uid, "PROD", num_conn, rate)
-            cons_resp_future = executor.submit(client_request, cons_stub, uid, "CONS", num_conn, rate)
-            ## SSH
-            prod_resp = prod_resp_future.result()
-            cons_resp = cons_resp_future.result()
-
-        print(prod_resp) ## Should this be printed?
-        prod_lstn = prod_resp.listeners
-        prod_app_lstn = prod_resp.prod_listeners
-        cons_lstn = cons_resp.listeners
-
-        print("Sending updated connection map information...")
-        print(uid)
-        update(prod_stub, uid, prod_resp.prod_listeners)
-        update(cons_stub, uid, prod_resp.listeners)
-
-@cli.command()
-@click.option('--num_conn', type=int, default=5)
-@click.option('--rate', type=int, default=10000)
-@click.option('--s2cs', default="localhost:5000")
-def request1(num_conn, rate, s2cs):
-    with grpc.insecure_channel(s2cs) as channel1:
-        prod_stub = scistream_pb2_grpc.ControlStub(channel1)
-        uid=str(uuid.uuid1())
-        with futures.ThreadPoolExecutor(max_workers=4) as executor:
-            click.echo(f"{uid} {s2cs} {utils.get_access_token()} PROD")
-            prod_resp_future = executor.submit(client_request, prod_stub, uid, "PROD", num_conn, rate)
-            prod_resp = prod_resp_future.result()
-            producer = producer_future.result()
-            click.echo(f"{uid} {s2cs} {utils.get_access_token()} PROD")
-
-        update(prod_stub, uid, prod_resp.prod_listeners)
-        print(uid)
-        click.echo(f"{uid} {s2cs} {utils.get_access_token()} CONS")
-
-@cli.command()
-@click.option('--num_conn', type=int, default=5)
-@click.option('--rate', type=int, default=10000)
-@click.option('--s2cs', default="localhost:5000")
-def request2(num_conn, rate, s2cs, access_code):
-    with grpc.insecure_channel(s2cs) as channel1:
-        prod_stub = scistream_pb2_grpc.ControlStub(channel1)
-        uid=str(uuid.uuid1())
-        with futures.ThreadPoolExecutor(max_workers=4) as executor:
-            prod_resp_future = executor.submit(client_request, prod_stub, uid, "PROD", num_conn, rate)
-            time.sleep(0.2)  # Possible race condition between REQ and HELLO
-            producer_future = executor.submit(IperfCtrl, uid, "PROD", s2cs)
-            prod_resp = prod_resp_future.result()
-            producer = producer_future.result()
-        update(prod_stub, uid, prod_resp.prod_listeners)
-        with futures.ThreadPoolExecutor(max_workers=4) as executor:
-            consumer_future = executor.submit(IperfCtrl, uid, "CONS", s2cs)
-            consumer = consumer_future.result()
-            ## APPctrl communicates with Scistream
-            ## Scistream tells it what port it should send the data to
+def release(uid, s2cs, metadata=None):
+    try:
+        with grpc.insecure_channel(s2cs) as channel:
+            stub = scistream_pb2_grpc.ControlStub(channel)
+            msg = scistream_pb2.Release(uid=uid)
+            resp = stub.release(msg, metadata=metadata)
+            print("Release completed")
+    except Exception as e:
+        print(f"Error during release: {e}")
 
 @cli.command()
 @click.option('--num_conn', type=int, default=5)
 @click.option('--rate', type=int, default=10000)
 @click.option('--s2cs', default="localhost:5000")
 @click.option('--mock', default=False)
-def prod_req(num_conn, rate, s2cs, mock):
+@click.option('--scope', default="")
+def prod_req(num_conn, rate, s2cs, mock, scope):
     with grpc.insecure_channel(s2cs) as channel:
         prod_stub = scistream_pb2_grpc.ControlStub(channel)
         uid = str(uuid.uuid1()) if not mock else "4f8583bc-a4d3-11ee-9fd6-034d1fcbd7c3"
         click.echo("uid; s2cs; access_token; role")
-        scope_id = get_scope_id(s2cs)
-        click.echo(f"{uid} {s2cs} {utils.get_access_token(scope_id)} PROD")
+        if scope == "":
+            scope = utils.get_scope_id(s2cs)
+        click.echo(f"{uid} {s2cs} {utils.get_access_token(scope)} PROD")
         with futures.ThreadPoolExecutor(max_workers=2) as executor:
             click.echo("waiting for hello message")
-            prod_resp_future = executor.submit(client_request, prod_stub, uid, "PROD", num_conn, rate, scope_id=scope_id)
+            prod_resp_future = executor.submit(client_request, prod_stub, uid, "PROD", num_conn, rate, scope_id=scope)
             prod_resp = prod_resp_future.result()
         print(prod_resp)  # Should this be printed?
         # Extracting listeners
         prod_lstn = prod_resp.listeners
         prod_app_lstn = prod_resp.prod_listeners
         # Update the prod_stub
-        update(prod_stub, uid, prod_resp.prod_listeners, scope_id=scope_id)
+        update(prod_stub, uid, prod_resp.prod_listeners, scope_id=scope)
         print(prod_resp.listeners)
 
 @cli.command()
 @click.option('--num_conn', type=int, default=5)
 @click.option('--rate', type=int, default=10000)
 @click.option('--s2cs', default="localhost:6000")
+@click.option('--scope', default="")
 @click.argument("uid")
 @click.argument("prod_lstn")
-def cons_req(num_conn, rate, s2cs, uid, prod_lstn):  # uid and prod_lstn are dependencies from PROD context
+def cons_req(num_conn, rate, s2cs, scope, uid, prod_lstn):  # uid and prod_lstn are dependencies from PROD context
     with grpc.insecure_channel(s2cs) as channel:
         cons_stub = scistream_pb2_grpc.ControlStub(channel)
-        scope_id = get_scope_id(s2cs)
-        click.echo(f"{uid} {s2cs} {utils.get_access_token(scope_id)} CONS")
+        if scope == "":
+            scope = utils.get_scope_id(s2cs)
+        click.echo(f"{uid} {s2cs} {utils.get_access_token(scope)} CONS")
         with futures.ThreadPoolExecutor(max_workers=2) as executor:
-            cons_resp_future = executor.submit(client_request, cons_stub, uid, "CONS", num_conn, rate, scope_id=scope_id)
+            cons_resp_future = executor.submit(client_request, cons_stub, uid, "CONS", num_conn, rate, scope_id=scope)
             cons_resp = cons_resp_future.result()
         cons_lstn = cons_resp.listeners
         # Update the cons_stub
@@ -263,11 +150,18 @@ def cons_req(num_conn, rate, s2cs, uid, prod_lstn):  # uid and prod_lstn are dep
             listener_array = prod_lstn.split(',')
         else:
             listener_array = [prod_lstn]
-        update(cons_stub, uid, listener_array, scope_id=scope_idexit)  # prod_lstn is a dependency from PROD context
+        update(cons_stub, uid, listener_array, scope_id=scope)
+        # prod_lstn is a dependency from PROD context
 
 
 @utils.authorize
 def client_request(stub, uid, role, num_conn, rate, scope_id="", metadata=None):
+    """
+    This behaves slightly different than release,
+    release gets an IP:port tuple as input
+    This receives the grpc stub
+    Not sure what are the implications
+    """
     try:
         print("started client request")
         request = scistream_pb2.Request(uid=uid, role=role, num_conn=num_conn, rate=rate)
@@ -282,6 +176,7 @@ def client_request(stub, uid, role, num_conn, rate, scope_id="", metadata=None):
 
 @utils.authorize
 def update(stub, uid, remote_listeners, scope_id="", metadata=None):
+    """This behaves very similar to client_request"""
     try:
         update_request = scistream_pb2.UpdateTargets(uid=uid, remote_listeners=remote_listeners)
         stub.update(update_request, metadata=metadata)